Rename castling module
